Replace `bin/check-diff` with a one-liner
This removes the indirection of the bash script so that it's immediately
clear from the github action definition file what is being checked.

It also checks for _any_ delta in the git repo, not just the `dist/`
directory. Any change should fail CI until it's either committed or
added to `.gitignore`.